Added new document to try an idea I had.
One of the trickiest problems that the program seems to have is finding the switchpoint. I took one step towards fixing this by using a uniform prior. However, sometimes it would find a point that might be better but is too late. I implemented a part of the analysis I did for my RCB stars. Where I find the average slope over a portion of the graph to find the lowest slope. I then give that point to PyMC3 as a testval so it tests around that area more. It seems to allow the program to work with only 20 points on the entire light curve.